added map_core_schema json files and their unit tests
