+++ crates/json_schema_validation/tests/fixtures/map_core_schema/map-keyrule-types.json
@@ -1,0 +1,229 @@
+{
+  "meta": {
+    "map_schema_version": "0.0.1",
+    "description": "Refined MAP Key Rule Types — single-level representation with explicit parameterization and schema grouping",
+    "generator": "MAP Types and Schema Refactor — 2025-07-19"
+  },
+  "holons": [
+    {
+      "key": "MAP Key Rule Schema",
+      "type": "#Schema",
+      "properties": {
+        "type_name": "MAP Key Rule Schema",
+        "type_name_plural": "MAP Key Rule Schemas",
+        "display_name": "MAP Key Rule Schema",
+        "display_name_plural": "MAP Key Rule Schemas",
+        "description": "Groups all type descriptors related to holon key derivation strategies in MAP."
+      }
+    },
+    {
+      "key": "KeyRuleType",
+      "type": "#TypeDescriptor",
+      "properties": {
+        "type_name": "KeyRuleType",
+        "type_name_plural": "KeyRuleTypes",
+        "display_name": "Abstract Key Rule Type",
+        "display_name_plural": "Abstract Key Rule Types",
+        "description": "Abstract holon type describing how keys are derived for other holons. Subtypes such as Format.KeyRuleType and Relationship.KeyRuleType define specific key derivation strategies. This type should not be instantiated directly, but it SHOULD be ExtendedBy the various specific KeyRuleTypes.",
+        "type_kind": "Holon"
+      },
+      "relationships": [
+        {
+          "name": "Extends",
+          "target": {
+            "$ref": "#MetaHolonType"
+          }
+        },
+        {
+          "name": "IsComponentOf",
+          "target": {
+            "$ref": "#MAP Key Rule Schema"
+          }
+        }
+      ]
+    },
+    {
+      "key": "Format.KeyRuleType",
+      "type": "#TypeDescriptor",
+      "properties": {
+        "type_name": "Format.KeyRuleType",
+        "type_name_plural": "Format.KeyRuleTypes",
+        "display_name": "Format Key Rule Type",
+        "display_name_plural": "Format Key Rule Types",
+        "description": "Derives a key from a format string and a list of properties. Each instance of this type defines a concrete rule used by a HolonType.",
+        "type_kind": "Holon"
+      },
+      "relationships": [
+        {
+          "name": "Extends",
+          "target": {
+            "$ref": "#KeyRuleType"
+          }
+        },
+        {
+          "name": "UsesKeyRule",
+          "target": {
+            "$ref": "#ExtendedType.KeyRuleType"
+          }
+        },
+        {
+          "name": "InstanceProperties",
+          "target": [
+            {
+              "$ref": "#MapFormatStringPropertyType"
+            }
+          ]
+        },
+        {
+          "name": "InstanceRelationships",
+          "target": [
+            {
+              "$ref": "#(FormatKeyRuleType)-[PropertyList]->(PropertyType)"
+            }
+          ]
+        },
+        {
+          "name": "IsComponentOf",
+          "target": {
+            "$ref": "#MAP Key Rule Schema"
+          }
+        },
+        {
+          "name": "Instances",
+          "target": [
+            {
+              "key": "TypeName.KeyRule",
+              "type": "#Format.KeyRuleType",
+              "properties": {
+                "format": "$0"
+              },
+              "relationships": [
+                {
+                  "name": "PropertyList",
+                  "target": [
+                    {
+                      "$ref": "#MapTypeNamePropertyType"
+                    }
+                  ]
+                }
+              ]
+            }
+          ]
+        }
+      ]
+    },
+    {
+      "key": "Relationship.KeyRuleType",
+      "type": "#TypeDescriptor",
+      "properties": {
+        "type_name": "Relationship.KeyRuleType",
+        "type_name_plural": "Relationship.KeyRuleTypes",
+        "display_name": "Relationship Key Rule Type",
+        "display_name_plural": "Relationship Key Rule Types",
+        "description": "Hardcoded rule: ($0)-[$1]->($2) using source_type, type_name, and target_type.",
+        "type_kind": "Holon"
+      },
+      "relationships": [
+        {
+          "name": "Extends",
+          "target": {
+            "$ref": "#KeyRuleType"
+          }
+        },
+        {
+          "name": "UsesKeyRule",
+          "target": {
+            "$ref": "#ExtendedType.KeyRuleType"
+          }
+        },
+        {
+          "name": "IsComponentOf",
+          "target": {
+            "$ref": "#MAP Key Rule Schema"
+          }
+        }
+      ]
+    },
+    {
+      "key": "ExtendedType.KeyRuleType",
+      "type": "#TypeDescriptor",
+      "properties": {
+        "type_name": "ExtendedType.KeyRuleType",
+        "type_name_plural": "ExtendedType.KeyRuleTypes",
+        "display_name": "Extended Type Key Rule Type",
+        "display_name_plural": "Extended Type Key Rule Types",
+        "description": "Hardcoded rule: <type_name>.<extends.key>",
+        "type_kind": "Holon"
+      },
+      "relationships": [
+        {
+          "name": "Extends",
+          "target": {
+            "$ref": "#KeyRuleType"
+          }
+        },
+        {
+          "name": "UsesKeyRule",
+          "target": {
+            "$ref": "#ExtendedType.KeyRuleType"
+          }
+        },
+        {
+          "name": "IsComponentOf",
+          "target": {
+            "$ref": "#MAP Key Rule Schema"
+          }
+        }
+      ]
+    },
+    {
+      "key": "None.KeyRuleType",
+      "type": "#TypeDescriptor",
+      "properties": {
+        "type_name": "None.KeyRuleType",
+        "type_name_plural": "None.KeyRuleTypes",
+        "display_name": "None Key Rule Type",
+        "display_name_plural": "None Key Rule Types",
+        "description": "Indicates a keyless holon. Must be embedded and never referenced.",
+        "type_kind": "Holon"
+      },
+      "relationships": [
+        {
+          "name": "Extends",
+          "target": {
+            "$ref": "#KeyRuleType"
+          }
+        },
+        {
+          "name": "UsesKeyRule",
+          "target": {
+            "$ref": "#ExtendedType.KeyRuleType"
+          }
+        },
+        {
+          "name": "IsComponentOf",
+          "target": {
+            "$ref": "#MAP Key Rule Schema"
+          }
+        }
+      ]
+    },
+    {
+      "key": "TypeName.KeyRule",
+      "type": "#Format.KeyRuleType",
+      "properties": {
+        "template_string": "$0"
+      },
+      "relationships": [
+        {
+          "name": "PropertyList",
+          "target": [
+            {
+              "$ref": "#MapTypeNamePropertyType"
+            }
+          ]
+        }
+      ]
+    }
+  ]
+}
